+++ app/android/app/src/main/kotlin/com/friend/ios/phonecalls/PhoneCallsPlugin.kt
@@ -1,4 +1,6 @@
-package com.friend.ios
+package com.friend.ios.phonecalls
+
+import com.friend.ios.*
 
 import android.app.Activity
 import android.content.Context
